requirements.txt
# Django & DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7
redis==4.6.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-cryptography==1.1
python-decouple==3.8

# Tasks & Background Jobs
celery==5.3.4
django-celery-beat==2.5.0
flower==2.0.1

# AI & Machine Learning
openai==1.3.5
langchain==0.0.335
chromadb==0.4.17

# Utilities
Pillow==10.1.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
urllib3==2.0.7

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==19.12.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Documentation
drf-spectacular==0.26.5

# Monitoring & Logging
sentry-sdk==1.38.0
django-extensions==3.2.3

# File Storage
django-storages==1.14.2
boto3==1.34.0

# Email
django-anymail==10.2

# Caching
django-redis==5.4.0 